Moved compose_magnetlink() to GUI utilities
