+++ Tribler/Test/Core/Utilities/test_utilities.py
@@ -6,8 +6,7 @@ from twisted.web.server import Site
 from twisted.web.util import Redirect
 
 from Tribler.Core.Utilities.network_utils import get_random_port
-from Tribler.Core.Utilities.utilities import compose_magnetlink, http_get, is_simple_match_query, is_valid_url, \
-    parse_magnetlink
+from Tribler.Core.Utilities.utilities import http_get, is_simple_match_query, is_valid_url, parse_magnetlink
 from Tribler.Test.test_as_server import AbstractServer
 from Tribler.Test.tools import trial_timeout
 
@@ -83,24 +82,3 @@ class TestMakeTorrent(AbstractServer):
 
         query2 = '"\xc1ubuntu"* OR "debian"*'
         self.assertFalse(is_simple_match_query(query2))
-
-    def test_compose_magnetlink(self):
-        infohash = "DC4B96CF85A85CEEDB8ADC4B96CF85A85CEEDB8A"
-        name = "Some torrent name"
-        trackers = ['http://tracker1.example.com:8080/announce', 'http://tracker1.example.com:8080/announce']
-
-        expected_link0 = ""
-        expected_link1 = "magnet:?xt=urn:btih:DC4B96CF85A85CEEDB8ADC4B96CF85A85CEEDB8A"
-        expected_link2 = "magnet:?xt=urn:btih:DC4B96CF85A85CEEDB8ADC4B96CF85A85CEEDB8A&dn=Some+torrent+name"
-        expected_link3 = "magnet:?xt=urn:btih:DC4B96CF85A85CEEDB8ADC4B96CF85A85CEEDB8A&dn=Some+torrent+name" \
-                         "&tr=http://tracker1.example.com:8080/announce&tr=http://tracker1.example.com:8080/announce"
-
-        composed_link0 = compose_magnetlink(None)
-        composed_link1 = compose_magnetlink(infohash)
-        composed_link2 = compose_magnetlink(infohash, name=name)
-        composed_link3 = compose_magnetlink(infohash, name=name, trackers=trackers)
-
-        self.assertEqual(composed_link0, expected_link0)
-        self.assertEqual(composed_link1, expected_link1)
-        self.assertEqual(composed_link2, expected_link2)
-        self.assertEqual(composed_link3, expected_link3)
