+++ Tribler/Test/GUI/test_util.py
@@ -4,7 +4,7 @@ import unittest
 
 from six.moves.urllib.parse import unquote_plus
 
-from TriblerGUI.utilities import quote_plus_unicode, unicode_quoter
+from TriblerGUI.utilities import compose_magnetlink, quote_plus_unicode, unicode_quoter
 
 
 class TestGUIUtilities(unittest.TestCase):
@@ -81,3 +81,24 @@ class TestGUIUtilities(unittest.TestCase):
 
         self.assertNotEqual(s, encoded)
         self.assertEqual(s, unquote_plus(encoded))
+
+    def test_compose_magnetlink(self):
+        infohash = "DC4B96CF85A85CEEDB8ADC4B96CF85A85CEEDB8A"
+        name = "Some torrent name"
+        trackers = ['http://tracker1.example.com:8080/announce', 'http://tracker1.example.com:8080/announce']
+
+        expected_link0 = ""
+        expected_link1 = "magnet:?xt=urn:btih:DC4B96CF85A85CEEDB8ADC4B96CF85A85CEEDB8A"
+        expected_link2 = "magnet:?xt=urn:btih:DC4B96CF85A85CEEDB8ADC4B96CF85A85CEEDB8A&dn=Some+torrent+name"
+        expected_link3 = "magnet:?xt=urn:btih:DC4B96CF85A85CEEDB8ADC4B96CF85A85CEEDB8A&dn=Some+torrent+name" \
+                         "&tr=http://tracker1.example.com:8080/announce&tr=http://tracker1.example.com:8080/announce"
+
+        composed_link0 = compose_magnetlink(None)
+        composed_link1 = compose_magnetlink(infohash)
+        composed_link2 = compose_magnetlink(infohash, name=name)
+        composed_link3 = compose_magnetlink(infohash, name=name, trackers=trackers)
+
+        self.assertEqual(composed_link0, expected_link0)
+        self.assertEqual(composed_link1, expected_link1)
+        self.assertEqual(composed_link2, expected_link2)
+        self.assertEqual(composed_link3, expected_link3)
