+++ TriblerGUI/widgets/downloadsdetailstabwidget.py
@@ -3,12 +3,10 @@ from __future__ import absolute_import
 from PyQt5.QtCore import Qt
 from PyQt5.QtWidgets import QAction, QTabWidget, QTreeWidgetItem
 
-from Tribler.Core.Utilities.utilities import compose_magnetlink
-
 from TriblerGUI.defs import *
 from TriblerGUI.tribler_action_menu import TriblerActionMenu
 from TriblerGUI.tribler_request_manager import TriblerRequestManager
-from TriblerGUI.utilities import copy_to_clipboard, format_size, format_speed, is_video_file
+from TriblerGUI.utilities import compose_magnetlink, copy_to_clipboard, format_size, format_speed, is_video_file
 from TriblerGUI.widgets.downloadfilewidgetitem import DownloadFileWidgetItem
